Fix WT^ update / MT hash calculation bug
* Fix a bug where if a user has set custom WT^ votes, but none of them
are valid, the miner will generate a block that has an MT hash equal to
the last one. This causes WT^ blocks remaining to stay the same as the
default vote was never applied.

TODO make SCDB reject a block where the MT hash is equal to the current
one.