Testing fix for core clientmodel block update bug
* Instead of ignoring and not signalling a block update in the Qt GUI
clientmodel if it happened too soon after the last update, cache the
ignored update and emit it later.

* Keep track of the most recent ignored block update in the client model

* Run a timer until synchronized (once synchronized updates aren't
        ignored regardless of how quickly they happen)

* When the timer expires, emit the most recent previously ignorred block
update

These changes fix a bitcoin core bug that causes the GUI block count
(and any models that update based on block count like our sidechain
 tables) to be invalid after reindexing. The block count isn't updated
to the actual chain tip as the signals are supressed for happening too
quickly. Another way to fix this is to simply delete the check for how
recent the last block update signal was. That requires deleting only a
few lines of code instead of adding more code like this commit. I have
tested both options and couldn't measure any performance differences but
have decided to maintain the limits of BlockTipChanged in the
clientmodel class just in case it could cause GUI issues on some systems
due to the large number of signals that would be created.