Skip invalid "deposit" instead of returning false
The updated deposit validation code is too good and is detecting WT^
payouts as invalid deposits. The previous behavior was to skip all
invalid deposits. The only thing invalid about the WT^ payout is the
destination script. We plan to remove validation of destination scripts
to allow sidechains to use any kind of deposit script anyway so skipping
the deposits with invalid destination scripts for now isn't perfect but
better than not using the new deposit validation code in the mempool &
miner.